🧪 Testing multiple line commits 🧪
This is the first test in which more verbosity might be added to
the commit messages in the future.  For this particular test the
global git editor was first modified to Sublime Text with:
git config --global core.editor "subl -n -w"
then invoking the editor with:
git commit --allow-empty -S
This will also be a real-time test of the way in which git
interprets double-quotes in the editor during the commit.